7_label_compliance/requirements.txt
langchain
langchain-openai
streamlit
python-dotenv
pillow
